Domain.Converter.Instances

Description

This module defines instances for converting Neo4j elements (nodes and relationships) into Haskell types and their corresponding interactants within the domain model.

Base

Instances of the FromValue type class to handle conversion from Bolt's Value type to various Haskell types such as Int, String, and Float, as well as lists and optional values.

ElemInteractant instances for translating Neo4j graph elements into domain-specific types like Molecule, Catalyst, Reaction, and more. The Identity type is introduced to facilitate the mapping of Neo4j objects and building relationships.

Instances of NodeMask and RelMask to prepare data for creating or updating nodes and relationships in the database.

This module is integral for enabling seamless interaction with Neo4j, facilitating the conversion of graph structures into Haskell types for further processing and manipulation.

Orphan instances

ElemInteractant PathMask
Converts a Neo4j path to a Haskell PathMask (Introduce PathMask instance to avoid introducing a new typeclass for a pseudo-collection of Interactant's).

Instance details

Methods

exactInteractant :: Elem -> Either ParsingError PathMask

FromValue String
Converts a Bolt Value of type T (Text) to a Haskell String.

Instance details

Methods

fromValue :: Value -> Either ParsingError String

maybeFromValue :: Maybe Value -> Maybe String

FromValue Float
Converts a Bolt Value of type F (Double) to a Haskell Float.

Instance details

Methods

fromValue :: Value -> Either ParsingError Float

maybeFromValue :: Maybe Value -> Maybe Float

FromValue Int
Converts a Bolt Value of type I (Int) to a Haskell Int.

Instance details

Methods

fromValue :: Value -> Either ParsingError Int

maybeFromValue :: Maybe Value -> Maybe Int

InteractantElem NodeMask
Converts a node interactant to a NodeMask.

Instance details

Methods

exactElem :: Interactant -> Either ParsingError NodeMask

InteractantElem RelMask
Converts a relation of Interactants to a RelMask.

Instance details

Methods

exactElem :: Interactant -> Either ParsingError RelMask

FromValue a => FromValue (Maybe a)
Converts a Bolt Value of type N (Null) to a Haskell Maybe type.

Instance details

Methods

fromValue :: Value -> Either ParsingError (Maybe a)

maybeFromValue :: Maybe Value -> Maybe (Maybe a)

FromValue a => FromValue [a]
Converts a Bolt Value of type L (List) to a Haskell list of the specified type.

Instance details

Methods

fromValue :: Value -> Either ParsingError [a]

maybeFromValue :: Maybe Value -> Maybe [a]

ElemInteractant (ACCELERATE, Identity)
Converts a Neo4j relationship of type ACCELERATE to a Haskell ACCELERATE and its Identity.

Instance details

Methods

exactInteractant :: Elem -> Either ParsingError (ACCELERATE, Identity)

ElemInteractant (Catalyst, Identity)
Converts a Neo4j node with label Catalyst to a Haskell Catalyst and its Identity.

Instance details

Methods

exactInteractant :: Elem -> Either ParsingError (Catalyst, Identity)

ElemInteractant (Molecule, Identity)
Converts a Neo4j node with label Molecule to a Haskell Molecule and its Identity.

Instance details

Methods

exactInteractant :: Elem -> Either ParsingError (Molecule, Identity)

ElemInteractant (PRODUCT_FROM, Identity)
Converts a Neo4j relationship of type PRODUCT_FROM to a Haskell PRODUCT_FROM and its Identity.

Instance details

Methods

exactInteractant :: Elem -> Either ParsingError (PRODUCT_FROM, Identity)

ElemInteractant (REAGENT_IN, Identity)
Converts a Neo4j relationship of type REAGENT_IN to a Haskell REAGENT_IN and its Identity.

Instance details

Methods

exactInteractant :: Elem -> Either ParsingError (REAGENT_IN, Identity)

ElemInteractant (Reaction, Identity)
Converts a Neo4j node with label Reaction to a Haskell Reaction and its Identity.

Instance details

Methods

exactInteractant :: Elem -> Either ParsingError (Reaction, Identity)

ElemInteractant (Interactant, Identity)
Converts a Neo4j node to a Haskell Interactant based on its label (if Node) or type (if URelationship).

Instance details

Methods

exactInteractant :: Elem -> Either ParsingError (Interactant, Identity)

ElemInteractant (FOLLOW, Identity)
Converts a Neo4j relationship of type FOLLOW to a Haskell FOLLOW and its Identity.

Instance details

Methods

exactInteractant :: Elem -> Either ParsingError (FOLLOW, Identity)

ElemInteractant (Mechanism, Identity)
Converts a Neo4j node with label Mechanism to a Haskell Mechanism and its Identity.

Instance details

Methods

exactInteractant :: Elem -> Either ParsingError (Mechanism, Identity)

ElemInteractant (Stage, Identity)
Converts a Neo4j node with label Stage to a Haskell Stage and its Identity.

Instance details

Methods

exactInteractant :: Elem -> Either ParsingError (Stage, Identity)

ElemInteractant (INCLUDE, Identity, Identity)
Converts a Neo4j relationship of type INCLUDE to a Haskell INCLUDE and its two Identity values.

Instance details

Methods

exactInteractant :: Elem -> Either ParsingError (INCLUDE, Identity, Identity)