Models.Chemical

Description

This module defines data types for chemical entities, including Molecule, Reaction, Catalyst, and their respective properties. It facilitates JSON serialization and deserialization, providing a structure for representing reactions and their components.

Documentation

type ReactionID = Int

type MoleculeID = Int

type CatalystID = Int

data Molecule

Constructors

Molecule

Fields

moleculeId :: MoleculeID

moleculeSmiles :: String

moleculeIupacName :: String

data Reaction

Constructors

Reaction

Fields

reactionId :: ReactionID

reactionName :: String

data Catalyst

Constructors

Catalyst

Fields

catalystId :: CatalystID

catalystSmiles :: String

catalystName :: Maybe String

newtype PRODUCT_FROM

Constructors

PRODUCT_FROM

Fields

productAmount :: Float

data ACCELERATE

def - the default value for ACCELERATE corresponds to Standard Temperature and Pressure (STP): - temperature = 273.15 K (Kelvin) - pressure = 101.325 kPa (kilopascals)

Constructors

ACCELERATE

Fields

temperature :: [Float]

pressure :: [Float]

newtype REAGENT_IN

Constructors

REAGENT_IN

Fields

reagentAmount :: Float