Models.Process

Description

This module defines the ReactionDetails, MechanismDetails, and ProcessDetails data types for representing the details of chemical reactions and their associated mechanisms.

The ReactionDetails type encapsulates information about a chemical reaction, including the reaction itself, inbound reagents (as pairs of REAGENT_IN and Molecule), outbound products (as pairs of PRODUCT_FROM and Molecule), and relevant conditions (as pairs of ACCELERATE and Catalyst).

The MechanismDetails type models the context of a mechanism (including its description via FOLLOW) and the interactants associated with each stage of the mechanism.

Finally, the ProcessDetails type aggregates both ReactionDetails and MechanismDetails, providing a comprehensive overview of a chemical process.

Documentation

data ReactionDetails

Constructors

ReactionDetails

Fields

reaction :: Reaction

inboundReagents :: [(REAGENT_IN, Molecule)]

outboundProducts :: [(PRODUCT_FROM, Molecule)]

conditions :: [(ACCELERATE, Catalyst)]

data MechanismDetails

Constructors

MechanismDetails

Fields

mechanismContext :: (Mechanism, FOLLOW)

stageInteractants :: [(Stage, [Interactant])]

data ProcessDetails

Constructors

ProcessDetails

Fields

reactionDetails :: ReactionDetails

mechanismDetails :: MechanismDetails